Revert "Add support for qafire"
This reverts commit abf999dd52a2069a592e9ee7de840eb208718dde.

I realised there was an issue where if I change Circle to only build PRs, then the master branch would not be built. This wouldnt be a problem if we change the default branch to master, but I think until QA-Fire is proven reliable we should keep with develop(staging) -> master(production).